Add nox session for creating release commits
Adds a new nox session, `make_release`, which takes a version number as an
argument, updates the changelog, and pushes a release branch and tag for that
version based on the current commit. All of the interactions with git are
automated, but creating a PR to merge the release branch back to main requires a
manual step for now, as AFAIK there's not a nice way to do it without also
requiring GitHub credentials.

There are a few sanity checks in place to make sure this release process doesn't
do anything bad:
- It will refuse to do anything if the repo is dirty or has untracked files, as
  these changes could be lost or unintentionally included in the release.
- It fetches from `origin` and checks that the release tag/branch being created
  doesn't already exist. There is technically a narrow window where two people
  could make the same release simultaneously, but GitHub should reject the
  second push anyway.
- It will only allow releases when the current branch is `main`; prerelease
  versions skip this check at present, as we could potentially want to make dev
  releases on side branches. If we in future want to enable making releases from
  stable branches (to patch old versions), that should be straightforward to
  exclude from this check as well.